Moved `util` to `__init__`
PyInstaller fails to find the util file with the import method used. Moved the 'util' functions to the '__init__' file to fix this issue.

Since the methods are only intended to be used and accessed with-in the module, they will not be added to the type hint file for the '__init__' file.